requirements.txt
requests
scrapetube
discord.py
python-dotenv
mcstatus
pytz